add react-router-dom, page layout
