+++ src/actions/collectionAction.js
@@ -1,0 +1,12 @@
+
+
+
+
+
+export default function collectionAction() {
+    return (
+        <div>
+            
+        </div>
+    )
+}
