+++ src/interfaces/ICollection.tsx
@@ -1,0 +1,7 @@
+export interface ICollection {
+
+    id: string,
+    name: string,
+    active: boolean,
+
+}
