+++ src/views/Admin.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Admin() {
+    return (
+        <div>
+            Admin Screen
+        </div>
+    )
+}
